Storage System Update (#467)
* storage system improvements:
- storage implementations derive from base storage, LocalFileStorage used by default
- simplify storage init, use 'local' for default storage, init s3 storage if set as default
- default storage path grouping by date rather than by user
- default local storage moves committed warcs and indexes to `/data/storage/YYYY-MM-DD/`
- tests: add tests for local storage and s3 storage commit (s3 tests run only if s3 auth is set)
- s3: migrate to boto3
- add STORAGE_ROOT default env var (defaults to /data/storage if not set)

* storage work: api cleanup, pass user, coll, recording to upload_file()
handle 'close_file' pubsub message to actually close the file! (fix for windows issues?)

* recorder misc fixes:
- try sleep(0.1) after pubsub to allow close_file to run (for windows)
- redis decode_responses=True for recorder redis
- test fix updated_at to >= compare

* recorder pubsub: add missing subscribe

* storagecommitter: put move and delete ops back into main recorder app to ensure moving and deleting happens after local files have been closed

* storage commit:
- send message to delete committed file via handle_delete message in case not yet closed
- tests: factor common storagecommitter tests into base, and download warc test

* storage work: more refactoring
- combine replace temp to perm move special case with committing collection that becomes permanent
- readd MOVE_Q for recordings that will be committed
- tests: update tests, no rec replay after committing (recording is closed)

* move temp: update api to match frontend behavior, moveTemp instead of move-temp, toColl instead of to-coll

* storage commit:
- ensure nginx serves the new '/data/storage' path as well
- tests: update tests to check that index is still written to user dir
- test login migrate: add tests for user deletion, ensure everything is cleared

* test cleanup: remove test for cdx in user dir, test for final move only

* tests: fix test_cdxj_cache() load_test() override params
storagecommitter: delete cdxj file directly after committing, don't notify writers (not locked)

* storage tests: make S3 test bucket configurable via WR_TEST_S3_ROOT env var, remove stray print

+++ webrecorder/webrecorder/models/recording.py
@@ -29,7 +29,7 @@ class Recording(RedisUniqueComponent):
     WARC_KEY = 'r:{rec}:warc'
 
     DEL_Q = 'q:del:{target}'
-    MOVE_Q = 'q:mov:{target}'
+    MOVE_Q = 'q:move:{target}'
 
     OPEN_REC_TTL = 5400
 
@@ -92,7 +92,8 @@ class Recording(RedisUniqueComponent):
         return 'Recording on ' + created_at
 
     def delete_me(self):
-        self.delete_warcs()
+        if not self.delete_warcs():
+            return False
 
         return self.delete_object()
 
@@ -108,6 +109,7 @@ class Recording(RedisUniqueComponent):
             yield n, v
 
     def delete_warcs(self):
+        fail = False
         with redis_pipeline(self.redis) as pi:
             for n, v in self.iter_all_files(skip_index=False):
                 parts = urlsplit(v)
@@ -115,22 +117,25 @@ class Recording(RedisUniqueComponent):
                     target = parts.netloc
                 elif parts.scheme:
                     target = parts.scheme
+                    if '+' in target:
+                        target = target.split('+', 1)[0]
                 else:
-                    target = 'nginx'
+                    target = 'local'
 
-                pi.rpush(self.DEL_Q.format(target=target), v)
+                if target == 'local':
+                    if self.redis.publish('handle_delete', v) == 0:
+                        fail = True
+                else:
+                    pi.rpush(self.DEL_Q.format(target=target), v)
+
+        return not fail
 
     def move_warcs(self, to_user):
-        move = {}
-        move['hkey'] = self.WARC_KEY.format(rec=self.my_id)
-        move['to_user'] = to_user.name
+        data = {'rec': self.my_id,
+                'user': to_user.name
+               }
 
-        with redis_pipeline(self.redis) as pi:
-            for n, v in self.iter_all_files(skip_index=False):
-                move['from'] = v
-                move['name'] = n
-
-                pi.rpush(self.MOVE_Q.format(target='local'), json.dumps(move))
+        self.redis.rpush(self.MOVE_Q.format(target='local'), json.dumps(data))
 
     def _get_pagedata(self, pagedata):
         key = self.PAGE_KEY.format(rec=self.my_id)
